requirements.txt
appdirs==1.4.4
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.1
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.11
platformdirs==3.5.0
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
